OCPBUGS-44576: Address comments
