+++ frontend/public/components/about-modal.tsx
@@ -11,7 +11,7 @@ import { Trans, useTranslation } from 'react-i18next';
 import { useClusterVersion, BlueArrowCircleUpIcon, useCanClusterUpgrade } from '@console/shared';
 import { isLoadedDynamicPluginInfo } from '@console/plugin-sdk/src';
 import { useDynamicPluginInfo } from '@console/plugin-sdk/src/api/useDynamicPluginInfo';
-import { getBrandingDetails } from './masthead-branding';
+import { getBrandingDetails } from './utils/branding';
 import {
   ReleaseNotesLink,
   ServiceLevel,
