+++ frontend/public/components/app.jsx
@@ -13,10 +13,10 @@ import store, { applyReduxExtensions } from '../redux';
 import { useTranslation } from 'react-i18next';
 import { coFetchJSON, appInternalFetch } from '../co-fetch';
 import { detectFeatures } from '../actions/features';
-import { setFlag } from '../actions/flagActions';
+import { setFlag } from '../actions/flags';
 import AppContents from './app-contents';
 import { Masthead } from './masthead';
-import { getBrandingDetails } from './masthead-branding';
+import { getBrandingDetails } from './utils/branding';
 import { ConsoleNotifier } from './console-notifier';
 import { ConnectedNotificationDrawer } from './notification-drawer';
 import { Navigation } from '@console/app/src/components/nav';
