+++ frontend/public/components/cluster-settings/cluster-operator.tsx
@@ -40,7 +40,7 @@ import {
   YellowExclamationTriangleIcon,
 } from '@console/shared';
 import RelatedObjectsPage from './related-objects';
-import { ClusterVersionConditionsLink, UpdatingMessageText } from './cluster-settings-messages';
+import { ClusterVersionConditionsLink, UpdatingMessageText } from './cluster-status';
 
 export const clusterOperatorReference: K8sResourceKindReference = referenceForModel(
   ClusterOperatorModel,
