+++ frontend/public/components/masthead.jsx
@@ -12,7 +12,7 @@ import {
 import { BarsIcon } from '@patternfly/react-icons/dist/esm/icons/bars-icon';
 import { useNavigate } from 'react-router-dom-v5-compat';
 import { MastheadToolbar } from './masthead-toolbar';
-import { getBrandingDetails } from './masthead-branding';
+import { getBrandingDetails } from './utils/branding';
 
 export const Masthead = React.memo(({ isMastheadStacked, isNavOpen, onNavToggle }) => {
   const details = getBrandingDetails();
