+++ frontend/public/components/monitoring/receiver-forms/receiver-form-props.ts
@@ -1,0 +1,5 @@
+export type FormProps = {
+  globals: { [key: string]: any };
+  formValues: { [key: string]: any };
+  dispatchFormChange: Function;
+};
